Introduce types for GameState variants
First commit in a while, getting back into this codebase.
Started by modifying things with the intention of relying more heavily on lenses.
Accomplished this to some extent but introduced the larger change of adding WaitingState, InProgressState, and OverState.
These types allow for the restriction of functions which modify gamestate so we don't represent invalid operations.
For example by restricting the type of ready to WaitingState rather than GameState we now don't have the weird behavior of doing nothing when readying an in progress game.
This resulted in a lot of changes, some of which were introduced hastily in an effort to get things compiling again.
Next I intend to introduce helper methods for modifying the GameState only when it is waiting in progress or over depending on the case, and throwing errors otherwise.
This pattern emerged naturally but is not abstracted yet so the code is a little clunky.
After that its generating more lenses for types buried deep in the state.
Long term I hope to reduce the foot print of the GameState module by leveraging lenses throughout.