Represent ready users with a Map in Waiting state
Representing users and ready users as lists made determining when a user was present a separate task.
When updating, using typical list operations, a manual verification was necessary to ensure an unknown user resulted in an Exception instead of a zero update.
This also resulted in one less piece of data in WaitingState and shorter, more symmetric code in other places.
It is important to note this means the user list is also maintained via this single map, so Map.keys gives us what was users.
Very happy with this small change.

Unrelated to the bulk of this commit, an additional action on pressing `b` was added that runs a small Scrabble snippit which starts an example game.