Add MonadScrabble to avoid exposing MTL operations
I recently realized that because of the n^2 instances problem the ScrabbleT monad transformer would be annoying in practice.
Not to mention it is questionable anyone will ever use it - including myself.
Scrabble and ScrabbleIO should be sufficient for my use cases.
On top of this I realized that because of defining these types as an alias around ScrabbleT, and ScrabbleT deriving MTL classes, Scrabble expressions could access MTL operations like put and get.
This broke the otherwise clean and controlled API I want to expose so I decided to address it by introducing a MonadScrabble type class.
Following this approach Scrabble and ScrabbleIO may be defined as newtypes around ScrabbleT without exposing the MTL operations, or requiring involved modifications to ScrabbleT operations.
The way this was accomplished was by defining ScrabbleTs operations as its MonadScrabble instance, and deriving the MonadScrabble instances for Scrabble and ScrabbleIO.

This changeset probably should have been added outside of this branch as it is unrelated, but it crossed my mind while doing this work so here it lives.
Overall not a big change to the game logic, but a nice fix to a hole in the API.