Implement starting game
Change EnterState to WaitingStatus, adding new CantStart variant.
Add case for pressing enter in NotEntering state which attempts to start the game.
The game can only be started when there are between 2 and 4 players.
When starting is attempted and this is not true we enter the CantStart status.
Similar to Invalid Entry we display a notifiction in this case.
If we have the appropraite number of players we modify the existing Scrabble game state by starting the game.
This consists of adding each user from the user list, and then ready'ing them all.