Add PlaceStatus to track selecting position / tile
Add PlaceStatus ADT to Place InProgressStatus.
Currently PlaceStatus consists of SelectingPosition and SelectingTile.
This is now used to determine focus for the tilesList and determine when to send events to the tilesList.
Scrolling in the tilesList was implemented, though selecting from the list and actually placing is to do next.