Implement tile placement (and removal)
Move tile placements field from PlaceStatus to InProgressApp.
Pressing enter on an empty spot and then enter on a tile will now place the tile on the board.
Pressing enter on a tile with a placed letter on it will put the tile back in the tiles list.
Because of the move of the placements field placed tiles will persist even after exiting the placing mode.
Submitting placed tiles for scoring is not currently implemented, along with properly re-populating the tiles list after turn changes.
Another concern is displaying played tiles in a more visible color, as well as colors to distinguish un-submitted placed tiles from already submitted tiles.
Current thought is to add a submit word option to the menu.